[17주차] 테이블 수정 및 보스 체력 int로 변경

+++ TheLordServer/TheLordServer.MongoDB/CollectionData/BossData.cs
@@ -6,7 +6,7 @@ namespace TheLordServer.MongoDB.CollectionData
     public class BossData : BaseData
     {
         public int Index { get; set; }
-        public float HP { get; set; }
+        public int HP { get; set; }
         public DateTime CreateTime { get; set; }
 
         public BossData ( ) : base ( )
